minor API improvements
