+++ src/commons/constants.py
@@ -70,6 +70,7 @@ MAESTRO_RIGHTSIZER_APPLICATION_TYPE = 'RIGHTSIZER'
 MAESTRO_RIGHTSIZER_LICENSES_APPLICATION_TYPE = 'RIGHTSIZER_LICENSES'
 RIGHTSIZER_APPLICATION_TYPES = (MAESTRO_RIGHTSIZER_APPLICATION_TYPE,
                                 MAESTRO_RIGHTSIZER_LICENSES_APPLICATION_TYPE)
+APPLICATION_TENANTS_ALL = '*'
 INPUT_STORAGE_ATTR = 'input_storage'
 OUTPUT_STORAGE_ATTR = 'output_storage'
 ADDED_AT_ATTR = 'added_at'
